# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/address/query": {
      "get": {
        "operationId": "getQuery",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of records to skip",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of records to take",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "withDeleted",
            "required": false,
            "in": "query",
            "description": "Include soft deleted records",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "loadEagerRelations",
            "required": false,
            "in": "query",
            "description": "Load eager relations",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "transaction",
            "required": false,
            "in": "query",
            "description": "Use transaction",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "comment",
            "required": false,
            "in": "query",
            "description": "Query comment",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Address"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all Addresss",
        "tags": [
          "address"
        ]
      }
    },
    "/address/count": {
      "get": {
        "operationId": "getCount",
        "parameters": [
          {
            "name": "withDeleted",
            "required": false,
            "in": "query",
            "description": "Include soft deleted records",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "transaction",
            "required": false,
            "in": "query",
            "description": "Use transaction",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "comment",
            "required": false,
            "in": "query",
            "description": "Query comment",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Total count of entities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get count of Addresss",
        "tags": [
          "address"
        ]
      }
    },
    "/address/{id}": {
      "get": {
        "operationId": "getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Address"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one Address",
        "tags": [
          "address"
        ]
      },
      "patch": {
        "operationId": "updateOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Address"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update one Address",
        "tags": [
          "address"
        ]
      },
      "delete": {
        "operationId": "deleteOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete one Address",
        "tags": [
          "address"
        ]
      }
    },
...
    "/address/download/{filename}": {
      "get": {
        "operationId": "download",
        "parameters": [
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download file related to Address",
        "tags": [
          "address"
        ]
      }
    },
...
    "/card/query": {
      "get": {
        "operationId": "getQuery",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of records to skip",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of records to take",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "withDeleted",
            "required": false,
            "in": "query",
            "description": "Include soft deleted records",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "loadEagerRelations",
            "required": false,
            "in": "query",
            "description": "Load eager relations",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "transaction",
            "required": false,
            "in": "query",
            "description": "Use transaction",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "comment",
            "required": false,
            "in": "query",
            "description": "Query comment",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Card"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all Cards",
        "tags": [
          "card"
        ]
      }
    },
    "/card/count": {
      "get": {
        "operationId": "getCount",
        "parameters": [
          {
            "name": "withDeleted",
            "required": false,
            "in": "query",
            "description": "Include soft deleted records",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "transaction",
            "required": false,
            "in": "query",
            "description": "Use transaction",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "comment",
            "required": false,
            "in": "query",
            "description": "Query comment",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Total count of entities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get count of Cards",
        "tags": [
          "card"
        ]
      }
    },
    "/card/{id}": {
      "get": {
        "operationId": "getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one Card",
        "tags": [
          "card"
        ]
      },
      "patch": {
        "operationId": "updateOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Card"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update one Card",
        "tags": [
          "card"
        ]
      },
      "delete": {
        "operationId": "deleteOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete one Card",
        "tags": [
          "card"
        ]
      }
    },
...
    "/card/download/{filename}": {
      "get": {
        "operationId": "download",
        "parameters": [
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download file related to Card",
        "tags": [
          "card"
        ]
      }
    },
...
    "/competition/query": {
      "get": {
        "operationId": "getQuery",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of records to skip",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of records to take",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "withDeleted",
            "required": false,
            "in": "query",
            "description": "Include soft deleted records",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "loadEagerRelations",
            "required": false,
            "in": "query",
            "description": "Load eager relations",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "transaction",
            "required": false,
            "in": "query",
            "description": "Use transaction",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "comment",
            "required": false,
            "in": "query",
            "description": "Query comment",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Competition"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all Competitions",
        "tags": [
          "competition"
        ]
      }
    },
    "/competition/count": {
      "get": {
        "operationId": "getCount",
        "parameters": [
          {
            "name": "withDeleted",
            "required": false,
            "in": "query",
            "description": "Include soft deleted records",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "transaction",
            "required": false,
            "in": "query",
            "description": "Use transaction",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "comment",
            "required": false,
            "in": "query",
            "description": "Query comment",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Total count of entities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get count of Competitions",
        "tags": [
          "competition"
        ]
      }
    },
    "/competition/{id}": {
      "get": {
        "operationId": "getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Competition"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one Competition",
        "tags": [
          "competition"
        ]
      },
      "patch": {
        "operationId": "updateOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Competition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update one Competition",
        "tags": [
          "competition"
        ]
      },
      "delete": {
        "operationId": "deleteOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete one Competition",
        "tags": [
          "competition"
        ]
      }
    },
...
    "/competition/download/{filename}": {
      "get": {
        "operationId": "download",
        "parameters": [
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download file related to Competition",
        "tags": [
          "competition"
        ]
      }
    },
...
    "/fixture/query": {
      "get": {
        "operationId": "getQuery",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of records to skip",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of records to take",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "withDeleted",
            "required": false,
            "in": "query",
            "description": "Include soft deleted records",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "loadEagerRelations",
            "required": false,
            "in": "query",
            "description": "Load eager relations",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "transaction",
            "required": false,
            "in": "query",
            "description": "Use transaction",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "comment",
            "required": false,
            "in": "query",
            "description": "Query comment",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fixture"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all Fixtures",
        "tags": [
          "fixture"
        ]
      }
    },
    "/fixture/count": {
      "get": {
        "operationId": "getCount",
        "parameters": [
          {
            "name": "withDeleted",
            "required": false,
            "in": "query",
            "description": "Include soft deleted records",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "transaction",
            "required": false,
            "in": "query",
            "description": "Use transaction",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "comment",
            "required": false,
            "in": "query",
            "description": "Query comment",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Total count of entities",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get count of Fixtures",
        "tags": [
          "fixture"
        ]
      }
    },
    "/fixture/{id}": {
      "get": {
        "operationId": "getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Fixture"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one Fixture",
        "tags": [
          "fixture"
        ]
      },
      "patch": {
        "operationId": "updateOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Fixture"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update one Fixture",
        "tags": [
          "fixture"
        ]
      },
      "delete": {
        "operationId": "deleteOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete one Fixture",
        "tags": [
          "fixture"
        ]
      }
    },
...
    "/fixture/download/{filename}": {
      "get": {
        "operationId": "download",
        "parameters": [
          {
            "name": "filename",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Download file related to Fixture",
        "tags": [
          "fixture"
        ]
      }
    },
...
  },
  "info": {
    "title": "I Watch Football API",
    "description": "The API Docs for I Watch Football",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Promise": {
        "type": "object",
        "properties": {}
      },
      "Address": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "deletedAt": {
            "type": "string"
          },
          "address1": {
            "type": "string",
            "description": "Primary address line",
            "example": "123 Main Street"
          },
          "address2": {
            "type": "string",
            "description": "Secondary address line",
            "example": "Suite 200"
          },
          "townOrCity": {
            "type": "string",
            "description": "Town or city",
            "example": "London"
          },
          "postcode": {
            "type": "string",
            "description": "Postcode or ZIP code",
            "example": "SW1A 1AA"
          },
          "country": {
            "type": "string",
            "description": "Country name",
            "example": "United Kingdom"
          },
          "location": {
            "type": "string",
            "description": "Location description or coordinates",
            "example": "51.509865, -0.118092"
          },
          "stadiumId": {
            "type": "number",
            "description": "Associated stadium ID",
            "example": 5
          },
          "stadium": {
            "nullable": true,
            "description": "Associated stadium entity",
            "allOf": [
              {
                "$ref": "#/components/schemas/Promise"
              }
            ]
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}